Rest of DB implemented
DAOs methods implemented.

Entities changed.
Home: CurrentRoomProfile variable added
LastMeasurement:
RoomProfile: bool IsDefault added
SensorData:

+++ SmartHomeAPI/Entities/LastMeasurement.cs
@@ -3,12 +3,13 @@
     public class LastMeasurement
     {
         
-        public int LastMeasurementId { get; set; } // Primary key
+        public long LastMeasurementId { get; set; } // Primary key
         public string? Name { get; set; }
         public SensorData? LastDataMeasurement { get; set; }
         public string? DeviceEui { get; set; }
+        
 
-        public LastMeasurement(int lastMeasurementId, string? name, SensorData? lastDataMeasurement, string? deviceEui)
+        public LastMeasurement(long lastMeasurementId, string? name, SensorData? lastDataMeasurement, string? deviceEui )
         {
             LastMeasurementId = lastMeasurementId;
             Name = name;
